Endpoints

+++ src/main/java/com/gini/domain/dto/PartDto.java
@@ -3,14 +3,13 @@ package com.gini.domain.dto;
 import com.gini.domain.enums.Manufacturer;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.UUID;
 
 
 public record PartDto(
         UUID id,
         String partName,
-        BigInteger partCount,
+        Integer partCount,
         String partNumber,
         BigDecimal price,
         String currency,
